Hide contextual help link
This is a little bit wonky for us since it links prominently to edX's
documentation. While there is helpful information there, it also makes
assumptions about edX's deployment, mentions features we aren't using,
etc., so I think it's best to keep hiding this.

Also, this added header item causes the page to break more quickly on
narrower windows, so this mitigates somewhat.