Update email confirmation success message
The message had said you could view your new email address on the
dashboard, which is innaccurate; it's on the account settings page.

While we could have updated the link to point there instead,
we've deemed it more useful to send users to their dashboard,
while also mentioning the new email address in the message.